aider: Added code to import the main Svelte component and mount it on the `svelte-app` div.
# Aider chat conversation:

USER: I added these *read-write* files: frontend/src/lib/index.js

ASSISTANT: Great! Now that we have the `frontend/src/lib/index.js` file, we can set up the Svelte app to mount on the `svelte-app` div in the `svelte.html` template.

In a Svelte app, the main JavaScript file (in this case `index.js`) is responsible for importing the main Svelte component and mounting it on a div in the HTML.

Here's how you can update the `index.js` file to import the main Svelte component (let's call it `App.svelte`) and mount it on the `svelte-app` div:

```javascript
frontend/src/lib/index.js
<<<<<<< HEAD

=======
import App from '../App.svelte';

const app = new App({
  target: document.getElementById('svelte-app'),
});

export default app;
>>>>>>> updated
```

This change assumes that the main Svelte component is named `App.svelte` and is located in the `src` directory. If the component has a different name or is located elsewhere, you should update the import statement accordingly.

Please confirm if these are the changes you want to make, or if there are other changes you'd like to implement.