aider: Updated map page to center on user's current location if permission is granted, otherwise default to current coordinates.
# Aider chat conversation:

USER: I added these *read-write* files: manage.py

ASSISTANT: Ok.

USER: Please modify the map page to be centered on the user's current location. The user should be asked for permission to get their current location, and if it is denied, the default center should remain at the same coordinates we're using right now.

ASSISTANT: To achieve this, we need to modify the `frontend/src/routes/map/+page.svelte` file. We will use the Geolocation API to get the user's current location. If the user denies the permission or if the location is not available, we will use the default coordinates.

Here is the *edit block* for the changes:

```svelte
frontend/src/routes/map/+page.svelte
<<<<<<< HEAD
  let lat = 42.65698624597273;
  let lng = -73.75144231302086;
  let zoom = 9;

  function updateData() {
    	zoom = map.getZoom();
    	lng = map.getCenter().lng;
    	lat = map.getCenter().lat;
  }

	onMount(async () => {
		const initialState = { lng: lng, lat: lat, zoom: zoom };
=======
  let lat;
  let lng;
  let zoom = 9;

  function updateData() {
    zoom = map.getZoom();
    lng = map.getCenter().lng;
    lat = map.getCenter().lat;
  }

  onMount(async () => {
    if (navigator.geolocation) {
      navigator.geolocation.getCurrentPosition((position) => {
        lat = position.coords.latitude;
        lng = position.coords.longitude;
      }, () => {
        lat = 42.65698624597273;
        lng = -73.75144231302086;
      });
    } else {
      lat = 42.65698624597273;
      lng = -73.75144231302086;
    }

    const initialState = { lng: lng, lat: lat, zoom: zoom };
>>>>>>> updated
```

This change will ask the user for their current location when the map page is loaded. If the user denies the permission or if the location is not available, the map will be centered on the default coordinates.